Heat of Solution
Python program to calculate the "heat of solution" of a compound given
the compound's molar mass (molarMass), how much of the compound you
dissolved (gramsCompound), how much water you dissolved the compound in
(gramsWater), the temperature of the water before the reaction in
Kelvin/Celsius (initTemp), and the temperature of the water after the
reaction in Kelvin/Celsius (finalTemp).

PRO TIP: If you only have the change in temperature during the
reaction, set initTemp to 0.0 and finalTemp to whatever that change was.